+++ app/src/main/java/com/costular/atomtasks/data/tasks/DefaultTasksRepository.kt
@@ -2,6 +2,7 @@ package com.costular.atomtasks.data.tasks
 
 import com.costular.atomtasks.data.toDomain
 import com.costular.atomtasks.domain.model.Task
+import com.costular.atomtasks.domain.repository.TasksRepository
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import java.time.LocalDate
